pyproject.toml
[project]
name = "stalfreds-podcast"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.0",
    "feedparser>=6.0.11",
    "pandas>=2.3.1",
    "playwright>=1.54.0",
    "requests>=2.32.4",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
]
